Normalize javascript index data
