+++ app/templates/javascript/event.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
