+++ app/templates/javascript/index.hbs
@@ -1,0 +1,40 @@
+<section class="javascript-intro">
+  <div class="container">
+    <div class="row">
+      <h2 class="col-lg-8">
+        {{model.nextEvent.title}}
+      </h2>
+    </div>
+    <div class="row">
+      <p class="col-lg-5">
+        Keep up with the advancement of prominent open source frameworks, libraries, and browser standards by attending this online event.
+      </p>
+      <p class="col-lg-5 push-lg-1">
+        Core team members will discuss topics such as upcoming releases, recent milestones, and community initiatives.
+      </p>
+    </div>
+    <div class="next-event">
+      {{moment-format model.nextEvent.date 'MMMM Do YYYY'}} <br>
+      {{moment-format model.nextEvent.timezoneDate 'h:mm a zz'}}. Online! <br>
+      <a href="http://eepurl.com/cvWBM5" class="button action-button" target="_blank">
+        Register
+      </a>
+    </div>
+  </div>
+</section>
+
+<section class="javascript-events">
+  <div class="container">
+    <ul class="row">
+      {{#each model.speakers as |author|}}
+        <li class="col-lg-4 col-md-6 speaker-card">
+          <strong class="framework-pop {{author.frameworks}}">{{author.frameworks}}</strong>
+          <div class="avatar">
+            <img src="{{author.image}}" alt="{{author.name}}" />
+          </div>
+          <h3>{{author.name}}</h3>
+        </li>
+      {{/each}}
+    </ul>
+  </div>
+</section>
