+++ blog/2017-28-01-tdjs-event-1.md
@@ -1,6 +1,6 @@
 ---
 title: This.Javascript
-date: Sat Jan 28 2017 09:00:00 GMT-0600 (CST)
+date: Sat Jan 28 2017 09:00:00 GMT-0800 (PST)
 author: thisdot
 brand: this-dot-js
 video: http://eepurl.com/cvWBM5
