+++ blog/authors/ben-lesh.md
@@ -3,6 +3,6 @@ name: Ben Lesh
 twitter: benlesh
 github: https://github.com/blesh
 image: /media/people/ben-lesh.png
-featured: tdjs1
+featured: tdjs-event-1
 frameworks: rxjs
 ---
