+++ blog/authors/dan-abramov.md
@@ -3,6 +3,6 @@ name: Dan Abramov
 twitter: dan_abramov
 github: https://github.com/gaearon
 image: /media/people/dan-abramov.jpg
-featured: tdjs1
+featured: tdjs-event-1
 frameworks: react
 ---
