+++ blog/authors/evan-you.md
@@ -3,6 +3,6 @@ name: Evan You
 twitter: youyuxi
 github: https://github.com/yyx990803
 image: /media/people/evan-you.jpeg
-featured: tdjs1
+featured: tdjs-event-1
 frameworks: vue
 ---
