+++ blog/authors/matt-beale.md
@@ -3,6 +3,6 @@ name: Matthew Beale
 twitter: mixonic
 github: https://github.com/mixonic
 image: /media/people/matt-beale.png
-featured: tdjs1
+featured: tdjs-event-1
 frameworks: ember
 ---
