+++ blog/authors/misk-hevery.md
@@ -3,6 +3,6 @@ name: Miško Hevery
 twitter: mhevery
 github: https://github.com/robdodson
 image: /media/people/misko-hevery.jpg
-featured: tdjs1
+featured: tdjs-event-1
 frameworks: angular
 ---
