+++ blog/authors/rob-dodson.md
@@ -3,6 +3,6 @@ name: Rob Dodson
 twitter: rob_dodson
 github: https://github.com/robdodson
 image: /media/people/rob-dodson.jpeg
-featured: tdjs1
+featured: tdjs-event-1
 frameworks: polymer
 ---
